feat(presentation/placelist): create file
